New GUI - Refactor and add save profiles
